+++ src/app/components/pagination/Pagination.tsx
@@ -1,12 +1,14 @@
 import styles from './Pagination.module.scss';
 import cn from 'classnames'
 import { useMemo, memo, useEffect } from 'react'
-import { changeActivePage, getIDS } from '@/app/store/AppSlice'
+import { changeActivePage, clearIDS, getIDS } from '@/app/store/AppSlice'
 import { removeProducts } from '@/app/store/ProductSlice'
 import { useDispatch, useSelector } from 'react-redux';
 import { AppDispatch, RootState } from '@/app/store/store';
 
 const Pagination = ({ pageCount = 1, siblingCount = 1, }) => {
+    const productsList = useSelector((state: RootState) => state.productsList)
+    const application = useSelector((state: RootState) => state.application)
     const activePage = useSelector((state: RootState) => state.application.activePage)
     const action = useSelector((state: RootState) => state.application.action)
     const params = useSelector((state: RootState) => state.application.params)
@@ -14,35 +16,65 @@ const Pagination = ({ pageCount = 1, siblingCount = 1, }) => {
     const dispatch = useDispatch<AppDispatch>()
 
     useEffect(() => {
-        if (action === 'filter') {
-            const query: any = {}
-            if (params.brand) {
-                query["brand"] = params.brand
-            }
-            if (params.price) {
-                query["price"] = params.price
-            }
-            if (params.product) {
-                query["product"] = params.product
-            }
+        const getData = () => {
+            if (action === 'filter') {
+                const query: any = {}
+                if (params.brand) {
+                    query["brand"] = params.brand
+                }
+                if (params.price) {
+                    query["price"] = params.price
+                }
+                if (params.product) {
+                    query["product"] = params.product
+                }
 
-            dispatch(getIDS({
-                "action": action,
-                "params": query
-            }))
+                dispatch(getIDS({
+                    "action": action,
+                    "params": {
+                        limit: 50,
+                        offset: application.params.offset,
+                        ...query,
+                    }
+                }))
+            }
+            else {
+                dispatch(getIDS({
+                    action,
+                    params: {
+                        limit: 50,
+                        offset: application.params.offset,
+                    }
+                }))
+            }
         }
-        else {
-            dispatch(getIDS({
-                action,
-            }))
+        if (application.ids.length === 0) {
+            if (application.error.includes('нет товаров')) {
+                const time = 3000
+                setTimeout(() => {
+                    console.log('error')
+                }, time)
+            }
+            if (application.error.includes('Unexpected')) {
+                const time = 2000
+                setTimeout(() => {
+                    getData()
+                }, time)
+            }
+            getData()
         }
-    }, [action])
+    }, [action, activePage, params.brand, params.price, params.product])
 
     let pag = []
     const PageBtn = ({ page }: { page: number }) => {
         return <div
             onClick={() => {
-                dispatch(removeProducts())
+                if (productsList.products.length !== 0) {
+                    dispatch(removeProducts())
+                }
+                if (application.ids.length !== 0) {
+                    dispatch(clearIDS())
+                }
                 dispatch(changeActivePage(page))
             }
             }
